spike: `ProviderService` python examples
